Add unit tests and implement serialization for Request
Unit tests have been added for images and request construction as well as request serialization. Moreover, a new constructor for `Request` in `web_socket.rs` has been implemented for better initialization of objects. `Request` is now also serializable which enhances the sharing of this data structure across different parts of the application.